add bot

+++ src/main/java/com/example/adminpanel/DB/IDAO/IDAO.java
@@ -4,8 +4,8 @@ import java.util.Collection;
 
 public interface IDAO <T> {
     Collection<T> getAll();
-    T get(int id);
-    void delete(int id);
+    T get(Long id);
+    void delete(Long id);
     void create(T subscribe);
     void update(T subscribe);
 }
